move ptr wrapper to header file
